fix bug

+++ reinforcement learning/agent/image_state/ppg_clr.py
@@ -95,8 +95,8 @@ class AgentImageStatePPGClr(AgentPPG):
 
         for _ in range(self.ppo_epochs):       
             for images, states, actions, rewards, dones, next_images, next_states in dataloader: 
-                self._training_ppo(images.float().to(self.device), states.float().to(self.device), actions.float().to(self.device), 
-                    rewards.float().to(self.device), dones.float().to(self.device), next_images.float().to(self.device), next_states.float().to(self.device))
+                self._training_ppo(images.to(self.device), states.to(self.device), actions.to(self.device), 
+                    rewards.to(self.device), dones.to(self.device), next_images.to(self.device), next_states.to(self.device))
 
         images, states, _, _, _, _, _ = self.ppo_memory.get_all_items()
         self.aux_ppg_memory.save_all(images, states)
@@ -110,7 +110,7 @@ class AgentImageStatePPGClr(AgentPPG):
 
         for _ in range(self.aux_ppg_epochs):       
             for images, states in dataloader:
-                self._training_aux_ppg(images.float().to(self.device), states.float().to(self.device))
+                self._training_aux_ppg(images.to(self.device), states.to(self.device))
 
         images, _ = self.aux_ppg_memory.get_all_items()
         self.aux_clr_memory.save_all(images)
@@ -123,7 +123,7 @@ class AgentImageStatePPGClr(AgentPPG):
 
         for _ in range(self.aux_clr_epochs):
             for input_images, target_images in dataloader:
-                self._training_aux_clr(input_images.float().to(self.device), target_images.float().to(self.device))            
+                self._training_aux_clr(input_images.to(self.device), target_images.to(self.device))            
 
         self.aux_clr_memory.clear_memory()
